+++ src/renderer/src/scripts/utils/datasetStructure.js
@@ -349,7 +349,6 @@ export const createStandardizedDatasetStructure = (datasetStructure, datasetEnti
     const standardizedStructure = JSON.parse(JSON.stringify(window.datasetStructureJSONObj));
 
     useGlobalStore.setState({ datasetStructureJSONObj: standardizedStructure });
-    reRenderTreeView();
     // --- Step 7: Revert any global changes to window.datasetStructureJSONObj ---
     window.datasetStructureJSONObj = originalStructure;